added point/spring classes...
added in the point and spring headers/classes with basic variables and
methods.

Attempted to initialize and draw the basic spring as a line. Failed
hard. No idea what I'm doing wrong, but the program crashes somewhere
along the way of drawing the vector. I think it fails to access the
point positions.